Updated again...

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,14 +10,14 @@
  */
 size_t print_list(const list_t *h)
 {
-	int counter = 0;
+	size_t counter = 0;
 
 	while (h != NULL)
 	{
 		if (h->str == NULL)
 			printf("[0](nil)\n");
 		else
-			printf("[%u]%s\n", h->len, h->str);
+			printf("[%d]%s\n", h->len, h->str);
 		h = h->next;
 		counter += 1;
 	}
